Move report generator into warp package
