+++ Warp_sci.py
@@ -25,6 +25,7 @@ from warp.ccwaveshift import waveshift_oneorder, PySpecshift, waveshiftClip
 from warp.SNratio_estimate import snestimate
 from warp.PyContinuum import PyContinuum
 from warp.vac2air_spec import vac2air_spec
+from warp import report
 from warp.output_layout import (
     INTERMEDIATE_OBJ_1DSPEC_DIRNAMES,
     INTERMEDIATE_OBJ_2DSPEC_DIRNAMES,
@@ -37,7 +38,6 @@ from warp.output_layout import (
 from warp.plotframes import plot_all_frames_norm, plot_all_frames_flux, plot_all_frames_flux_BG, plot_2dimages_mask, \
     plot_2dimages, snr_plots, plot_combined_norm, plot_2dimages_sv, peak_count_fwhm, aperture_plot, cosmicRay2dImages
 from warp.badpixmask import pyfixpix, cosmicRayMask
-import tex_source_maker
 
 
 class shortFile(str):
@@ -1137,7 +1137,7 @@ def Warp_sci(listfile, rawdatapath, calibpath, destpath, viewerpath="INDEF", que
         remove_or_move(conf.status, "reduction_log", trashdir, 1)
 
     if not noreport:
-        tex_source_maker.tex_source_make(conf, fsr, logo)
+        report.tex_source_make(conf, fsr, logo, pipeline_ver=pipeline_ver)
 
     # remove trash directory
     constant_str_length("Finished.")
